modify test cases for only http

+++ qa/L0_multiple_ports/test.sh
@@ -34,44 +34,18 @@ DATADIR=`pwd`/models
 SERVER=/opt/tensorrtserver/bin/trtserver
 source ../common/util.sh
 
-SP_ARR=(8008 8009 8010 8011 -1 8004 8004 8004 8004 -1 -1 -1)
-HP_ARR=(8008 8010 8010 8009 8005 -1 8005 8005 -1 8004 -1 -1)
-PP_ARR=(8008 8011 8011 8010 8005 8005 -1 8006 -1 -1 8004 -1)
-IP_ARR=(8008 8011 8008 8008 8006 8006 8006 -1 -1 -1 -1 8004)
-len=${#SP_ARR[@]}
+HP_ARR=(8000 8005 -1)
+len=${#HP_ARR[@]}
 rm -f $CLIENT_LOG $SERVER_LOG
 
 RET=0
-# HTTP + GRPC w/o Interleaved
+# HTTP Health w/o Main Port
 for (( n=0; n<$len; n++ )) ; do
-  SERVER_ARGS_ADD_GRPC="--grpc-status-port ${SP_ARR[n]} --grpc-health-port ${HP_ARR[n]} \
-    --grpc-profile-port ${PP_ARR[n]} --grpc-infer-port ${IP_ARR[n]} --allow-grpc 1"
-  SERVER_ARGS="--model-store=$DATADIR $SERVER_ARGS_ADD_GRPC"
-
-  run_server_nowait
-  sleep 10
-  if [ "$SERVER_PID" == "0" ]; then
-      echo -e "\n***\n*** Failed to start $SERVER\n***"
-      cat $SERVER_LOG
-      exit 1
-  fi
-
-  set +e
-  python $MULTI_PORT_TESTS_PY -v >>$CLIENT_LOG 2>&1 -sp ${SP_ARR[n]} -hp ${HP_ARR[n]} -pp ${PP_ARR[n]} -ip ${IP_ARR[n]} -i grpc
-  if [ $? -ne 0 ]; then
-      RET=1
-  fi
-  set -e
-
-  kill $SERVER_PID
-  wait $SERVER_PID
-
-  SERVER_ARGS_ADD_HTTP="--http-status-port ${SP_ARR[n]} --http-health-port ${HP_ARR[n]} \
-    --http-profile-port ${PP_ARR[n]} --http-infer-port ${IP_ARR[n]} --allow-http 1"
+  SERVER_ARGS_ADD_HTTP="--http-health-port ${HP_ARR[n]} --allow-http 1"
   SERVER_ARGS="--model-store=$DATADIR $SERVER_ARGS_ADD_HTTP"
 
   run_server_nowait
-  sleep 10
+  sleep 5
   if [ "$SERVER_PID" == "0" ]; then
       echo -e "\n***\n*** Failed to start $SERVER\n***"
       cat $SERVER_LOG
@@ -79,7 +53,7 @@ for (( n=0; n<$len; n++ )) ; do
   fi
 
   set +e
-  python $MULTI_PORT_TESTS_PY -v >>$CLIENT_LOG 2>&1 -sp ${SP_ARR[n]} -hp ${HP_ARR[n]} -pp ${PP_ARR[n]} -ip ${IP_ARR[n]} -i http
+  python $MULTI_PORT_TESTS_PY -v >>$CLIENT_LOG 2>&1 -hp ${HP_ARR[n]} -i http
   if [ $? -ne 0 ]; then
       RET=1
   fi
@@ -89,70 +63,49 @@ for (( n=0; n<$len; n++ )) ; do
   wait $SERVER_PID
 done
 
-# HTTP + GRPC w/ Interleaved
-P=(8005 -1)
+# HTTP + Health
+P=(8004 -1)
 for (( i=0; i<2; i++ )) ; do
   for (( n=0; n<$len; n++ )) ; do
-    SERVER_ARGS_ADD_GRPC="--grpc-port ${P[i]} --grpc-status-port ${SP_ARR[n]} --grpc-health-port ${HP_ARR[n]} \
-      --grpc-profile-port ${PP_ARR[n]} --grpc-infer-port ${IP_ARR[n]} --allow-grpc 1"
-    SERVER_ARGS="--model-store=$DATADIR $SERVER_ARGS_ADD_GRPC"
-
-    run_server_nowait
-    sleep 10
-    if [ "$SERVER_PID" == "0" ]; then
-        echo -e "\n***\n*** Failed to start $SERVER\n***"
-        cat $SERVER_LOG
-        exit 1
-    fi
-
-    set +e
-    python $MULTI_PORT_TESTS_PY -v >>$CLIENT_LOG 2>&1 -p 8005 -sp ${SP_ARR[n]} -hp ${HP_ARR[n]} -pp ${PP_ARR[n]} -ip ${IP_ARR[n]} -i grpc
-    if [ $? -ne 0 ]; then
-        RET=1
-    fi
-    set -e
-
-    kill $SERVER_PID
-    wait $SERVER_PID
-
-    SERVER_ARGS_ADD_HTTP="--http-port ${P[i]} --http-status-port ${SP_ARR[n]} --http-health-port ${HP_ARR[n]} \
-      --http-profile-port ${PP_ARR[n]} --http-infer-port ${IP_ARR[n]} --allow-http 1"
+    SERVER_ARGS_ADD_HTTP="--http-port ${P[i]} --http-health-port ${HP_ARR[n]} \
+      --allow-http 1"
     SERVER_ARGS="--model-store=$DATADIR $SERVER_ARGS_ADD_HTTP"
 
     run_server_nowait
-    sleep 10
-    if [ "$SERVER_PID" == "0" ]; then
-        echo -e "\n***\n*** Failed to start $SERVER\n***"
-        cat $SERVER_LOG
-        exit 1
-    fi
+    sleep 5
+    if ! [[ "${P[i]}" == "-1" && "${HP_ARR[n]}" == "-1" ]] ; then
+        if [ "$SERVER_PID" == "0" ]; then
+            echo -e "\n***\n*** Failed to start $SERVER\n***"
+            cat $SERVER_LOG
+            exit 1
+        else
+          set +e
+          python $MULTI_PORT_TESTS_PY -v >>$CLIENT_LOG 2>&1 -p ${P[i]} -hp ${HP_ARR[n]} -i http
+          if [ $? -ne 0 ]; then
+              RET=1
+          fi
+          set -e
 
-    set +e
-    python $MULTI_PORT_TESTS_PY -v >>$CLIENT_LOG 2>&1 -p 8005 -sp ${SP_ARR[n]} -hp ${HP_ARR[n]} -pp ${PP_ARR[n]} -ip ${IP_ARR[n]}
-    if [ $? -ne 0 ]; then
-        RET=1
+          kill $SERVER_PID
+          wait $SERVER_PID
+        fi
     fi
-    set -e
-
-    kill $SERVER_PID
-    wait $SERVER_PID
   done
 done
 
 # CUSTOM CASES
-# set http ports to -1 after setting them to 8007
-SERVER_ARGS_ADD_HTTP="--http-status-port 8007 --http-health-port 8007 \
-  --http-profile-port 8007 --http-infer-port 8007 --http-port -1 --allow-http 1"
+# set http port to -1 after setting health to 8007
+SERVER_ARGS_ADD_HTTP="--http-health-port 8007 --http-port -1 --allow-http 1"
 SERVER_ARGS="--model-store=$DATADIR $SERVER_ARGS_ADD_HTTP"
 run_server_nowait
-sleep 10
-if [ "$SERVER_PID" == "0" ]; then
-    echo -e "\n***\n*** Failed to start $SERVER\n***"
+sleep 5
+if [ "$SERVER_PID" != "0" ]; then
+    echo -e "\n***\n*** Should not have started $SERVER\n***"
     cat $SERVER_LOG
     exit 1
 fi
 set +e
-python $MULTI_PORT_TESTS_PY -v >>$CLIENT_LOG 2>&1 -p -1 -sp 8007 -hp 8007 -pp 8007 -ip 8007
+python $MULTI_PORT_TESTS_PY -v >>$CLIENT_LOG 2>&1 -p -1 -hp 8007 -i http
 if [ $? -ne 0 ]; then
     RET=1
 fi
@@ -160,18 +113,17 @@ set -e
 kill $SERVER_PID
 wait $SERVER_PID
 # allow overrules - grpc still works
-SERVER_ARGS_ADD_HTTP="--http-status-port 8007 --http-health-port 8007 \
-  --http-profile-port 8007 --http-infer-port 8007 --http-port -1 --allow-http 0"
+SERVER_ARGS_ADD_HTTP="--http-port -1 --http-health-port 8007 --allow-http 0"
 SERVER_ARGS="--model-store=$DATADIR $SERVER_ARGS_ADD_HTTP"
 run_server_nowait
-sleep 10
+sleep 5
 if [ "$SERVER_PID" == "0" ]; then
     echo -e "\n***\n*** Failed to start $SERVER\n***"
     cat $SERVER_LOG
     exit 1
 fi
 set +e
-python $MULTI_PORT_TESTS_PY -v >>$CLIENT_LOG 2>&1 -p -1 -sp 8007 -hp 8007 -pp 8007 -ip 8007
+python $MULTI_PORT_TESTS_PY -v >>$CLIENT_LOG 2>&1 -p -1 -hp 8007 -i http
 if [ $? -ne 0 ]; then
     RET=1
 fi
@@ -179,28 +131,21 @@ set -e
 kill $SERVER_PID
 wait $SERVER_PID
 # overlap with grpc default
-SERVER_ARGS_ADD_HTTP="--http-status-port 8001 --http-health-port 8007 \
-  --http-profile-port 8007 --http-infer-port 8007"
+SERVER_ARGS_ADD_HTTP="--http-health-port 8001"
 SERVER_ARGS="--model-store=$DATADIR $SERVER_ARGS_ADD_HTTP"
 run_server_nowait
-sleep 10
+sleep 5
 if [ "$SERVER_PID" != "0" ]; then
     echo -e "\n***\n*** Should not have started $SERVER\n***"
     cat $SERVER_LOG
     exit 1
 fi
-set +e
-python $MULTI_PORT_TESTS_PY -v >>$CLIENT_LOG 2>&1 -sp 8001 -hp 8007 -pp 8007 -ip 8007
-if [ $? -ne 0 ]; then
-    RET=1
-fi
-set -e
 # overlap with metrics default
 SERVER_ARGS_ADD_HTTP="--http-status-port 8002 --http-health-port 8007 \
   --http-profile-port 8007 --http-infer-port 8007"
 SERVER_ARGS="--model-store=$DATADIR $SERVER_ARGS_ADD_HTTP"
 run_server_nowait
-sleep 10
+sleep 5
 if [ "$SERVER_PID" != "0" ]; then
     echo -e "\n***\n*** Should not have started $SERVER\n***"
     cat $SERVER_LOG
@@ -211,7 +156,7 @@ SERVER_ARGS_ADD_HTTP="--http-status-port 8002 --http-health-port 8007 \
   --http-profile-port 8007 --http-infer-port 8007 --allow_metrics 0"
 SERVER_ARGS="--model-store=$DATADIR $SERVER_ARGS_ADD_HTTP"
 run_server_nowait
-sleep 10
+sleep 5
 if [ "$SERVER_PID" == "0" ]; then
     echo -e "\n***\n*** Failed to start $SERVER\n***"
     cat $SERVER_LOG
